+++ api_tabular/app.py
@@ -4,10 +4,10 @@ import sentry_sdk
 from aiohttp import web, ClientSession
 
 from sentry_sdk.integrations.aiohttp import AioHttpIntegration
-from udata_hydra_csvapi import config
-from udata_hydra_csvapi.query import get_resource, get_resource_data
-from udata_hydra_csvapi.utils import build_sql_query_string, build_link_with_page
-from udata_hydra_csvapi.error import QueryException
+from api_tabular import config
+from api_tabular.query import get_resource, get_resource_data
+from api_tabular.utils import build_sql_query_string, build_link_with_page
+from api_tabular.error import QueryException
 
 routes = web.RouteTableDef()
